Fix problem with Netty buffer and writer-index
Netty-buffer is wired to work with explicit, absoute writers.
Which means it expects you to sequentially write fields, using
`write*` methods, instead of relative ones `set*`.

This leads to the problem of buffer copying / unwrapping underlying
direct or heap buffers, which, in case with `set*` methods is
using `writerIndex`, which is still 0, therefore returning an
empty buffer despite the fact it's filled up.

Since all the payloads are allocated statically and message size
is pre-set, we have to work around it by setting an explicit message
size.