Change 2 tests that were not testing anything useful
They were of the form (is (= expr1) expr2), which always passes,
because (= expr1) is always true.  Changed them to what was probably
intended, which is (is (= expr1 expr2)).

The tests fail now, but I do not know enough about the function being
tested to know whether it is behaving as desired and the test should
be changed, or whether the test is correct and the function behavior
should change.